+++ Classes/Domain/Repository/TeaRepository.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace TTN\Tea\Domain\Repository;
 
 use TTN\Tea\Domain\Model\Tea;
-use TTN\Tea\Domain\Repository\Traits\StoragePageAgnosticTrait;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
 use TYPO3\CMS\Extbase\Persistence\Repository;
@@ -15,8 +14,6 @@ use TYPO3\CMS\Extbase\Persistence\Repository;
  */
 class TeaRepository extends Repository
 {
-    use StoragePageAgnosticTrait;
-
     protected $defaultOrderings = ['title' => QueryInterface::ORDER_ASCENDING];
 
     /**
@@ -25,6 +22,7 @@ class TeaRepository extends Repository
     public function findByOwnerUid(int $ownerUid): QueryResultInterface
     {
         $query = $this->createQuery();
+        $query->setQuerySettings($query->getQuerySettings()->setRespectStoragePage(false));
         $query->matching($query->equals('ownerUid', $ownerUid));
 
         return $query->execute();